move viewModel to right package
